score-ai-backend/requirements.txt
Flask
Flask-Cors
pydantic
pydantic-settings
google-cloud-storage
google-cloud-firestore
firebase-admin
google-genai
gunicorn
PyMuPDF
flasgger
firebase-functions
